+++ src/components/atom/Loading/Loading.css
@@ -1,0 +1,14 @@
+.loading-deembackground {
+  width: 100vw;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.2);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.loadding-spinner {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
